requirements.txt
mcp>=1.0.0
pyyaml>=6.0
pytest>=7.0.0
